requirements.txt
# Python3.12
requests==2.32.3
python-telegram-bot==21.4